made dataset download only occur if there is no dataset folder

+++ historical_model/run_model.sh
@@ -1,6 +1,8 @@
 export XRT_TPU_CONFIG="localservice;0;localhost:51011"
 unset LD_PRELOAD
-gsutil cp gs://search-query-classification-us-central1-c/aol_input_data.zip .
-unzip -o aol_input_data.zip
-rm -f aol_input_data.zip
+if [ ! -d "datasets" ]; then
+  gsutil cp gs://search-query-classification-us-central1-c/aol_input_data.zip .
+  unzip -o aol_input_data.zip
+  rm -f aol_input_data.zip
+fi
 python3 run_model.py
